requirements.txt
Flask==3.0.0
openai==1.54.0
python-dotenv==1.0.0
httpx==0.27.0
requests==2.31.0
beautifulsoup4==4.12.2
spotipy==2.24.0
